+++ src/common/Error/handlers/ChunkLoadError.tsx
@@ -1,6 +1,7 @@
 import { Button } from 'oa-components'
 import { useEffect } from 'react'
 import { Box, Text } from 'theme-ui'
+import { attemptReload, isReloaded } from './reloader'
 
 /**
  * Handle screen-of-death issue where the user has tried to load JS code that no longer exists on the server
@@ -10,19 +11,10 @@ import { Box, Text } from 'theme-ui'
  * Adapted from https://mitchgavan.com/code-splitting-react-safely/
  */
 export const ChunkLoadErrorHandler = () => {
-  const isReloaded = getWithExpiry('error_reload') ? true : false
-
+  // when error first seen attempt to resolve with a reload
   useEffect(() => {
-    toggleReload()
+    attemptReload()
   })
-
-  /** attempt to reload, use localstorage utility methods to avoid infinite loops */
-  const toggleReload = () => {
-    if (!isReloaded) {
-      setWithExpiry('error_reload', 'true', 10000)
-      window.location.reload()
-    }
-  }
 
   return (
     <Box
@@ -35,12 +27,12 @@ export const ChunkLoadErrorHandler = () => {
         paddingBottom: '20vh',
       }}
     >
-      {isReloaded && (
+      {isReloaded() && (
         <>
           <Text mb={4} sx={{ fontSize: 3 }}>
             The website got some updates
           </Text>
-          <Button onClick={() => toggleReload()}>Load the new site.</Button>
+          <Button onClick={() => attemptReload()}>Load the new site.</Button>
           <Text mt={4} sx={{ fontSize: 0.75 }}>
             If it doesn't work you might need to close and re-open your browser
             <br></br>Or send us a{' '}
@@ -53,27 +45,4 @@ export const ChunkLoadErrorHandler = () => {
       )}
     </Box>
   )
-}
-
-function setWithExpiry(key: string, value: string, ttl: number) {
-  const item = {
-    value: value,
-    expiry: new Date().getTime() + ttl,
-  }
-  localStorage.setItem(key, JSON.stringify(item))
-}
-
-function getWithExpiry(key: string): string | null {
-  const itemString = window.localStorage.getItem(key)
-  if (!itemString) return null
-
-  const item = JSON.parse(itemString)
-  const isExpired = new Date().getTime() > item.expiry
-
-  if (isExpired) {
-    localStorage.removeItem(key)
-    return null
-  }
-
-  return item.value
 }
